move types declaration from clean-and-save to types
